+++ app/src/main/cpp/skyline/common.cpp
@@ -4,6 +4,7 @@
 #include <android/log.h>
 #include "common.h"
 #include "nce.h"
+#include "soc.h"
 #include "gpu.h"
 #include "audio.h"
 #include "input.h"
@@ -55,6 +56,7 @@ namespace skyline {
     DeviceState::DeviceState(kernel::OS *os, std::shared_ptr<JvmManager> jvmManager, std::shared_ptr<Settings> settings, std::shared_ptr<Logger> logger)
         : os(os), jvm(std::move(jvmManager)), settings(std::move(settings)), logger(std::move(logger)) {
         // We assign these later as they use the state in their constructor and we don't want null pointers
+        soc = std::make_shared<soc::SOC>(*this);
         gpu = std::make_shared<gpu::GPU>(*this);
         audio = std::make_shared<audio::Audio>(*this);
         nce = std::make_shared<nce::NCE>(*this);
