+++ app/src/main/cpp/skyline/gpu.h
@@ -3,30 +3,16 @@
 
 #pragma once
 
-#include "gpu/gpfifo.h"
-#include "gpu/syncpoint.h"
-#include "gpu/engines/maxwell_3d.h"
 #include "gpu/presentation_engine.h"
 
 namespace skyline::gpu {
     /**
-     * @brief A common interfaces to the GPU where all objects relevant to it are present
+     * @brief An interface to host GPU structures, anything concerning host GPU/Presentation APIs is encapsulated by this
      */
     class GPU {
-      private:
-        const DeviceState &state;
-
       public:
         PresentationEngine presentation;
-        vmm::MemoryManager memoryManager;
-        std::shared_ptr<engine::Engine> fermi2D;
-        std::shared_ptr<engine::Maxwell3D> maxwell3D;
-        std::shared_ptr<engine::Engine> maxwellCompute;
-        std::shared_ptr<engine::Engine> maxwellDma;
-        std::shared_ptr<engine::Engine> keplerMemory;
-        std::array<Syncpoint, constant::MaxHwSyncpointCount> syncpoints{};
-        gpfifo::GPFIFO gpfifo;
 
-        GPU(const DeviceState &state) : state(state), presentation(state), memoryManager(state), gpfifo(state), fermi2D(std::make_shared<engine::Engine>(state)), keplerMemory(std::make_shared<engine::Engine>(state)), maxwell3D(std::make_shared<engine::Maxwell3D>(state)), maxwellCompute(std::make_shared<engine::Engine>(state)), maxwellDma(std::make_shared<engine::Engine>(state)) {}
+        GPU(const DeviceState &state) : presentation(state) {}
     };
 }
