+++ app/src/main/cpp/skyline/soc/gmmu.cpp
@@ -2,10 +2,12 @@
 // Copyright © 2020 Skyline Team and Contributors (https://github.com/skyline-emu/)
 
 #include <kernel/types/KProcess.h>
-#include "memory_manager.h"
+#include "gmmu.h"
 
-namespace skyline::gpu::vmm {
-    MemoryManager::MemoryManager(const DeviceState &state) : state(state) {
+namespace skyline::soc::gmmu {
+    constexpr u64 GpuPageSize{1 << 16}; //!< The page size of the GPU address space
+
+    GraphicsMemoryManager::GraphicsMemoryManager(const DeviceState &state) : state(state) {
         constexpr u64 gpuAddressSpaceSize{1UL << 40}; //!< The size of the GPU address space
         constexpr u64 gpuAddressSpaceBase{0x100000}; //!< The base of the GPU address space - must be non-zero
 
@@ -14,9 +16,9 @@ namespace skyline::gpu::vmm {
         chunks.push_back(baseChunk);
     }
 
-    std::optional<ChunkDescriptor> MemoryManager::FindChunk(ChunkState desiredState, u64 size, u64 alignment) {
+    std::optional<ChunkDescriptor> GraphicsMemoryManager::FindChunk(ChunkState desiredState, u64 size, u64 alignment) {
         auto chunk{std::find_if(chunks.begin(), chunks.end(), [desiredState, size, alignment](const ChunkDescriptor &chunk) -> bool {
-            return (alignment ? util::IsAligned(chunk.virtAddr, alignment) : true) && chunk.size > size && chunk.state == desiredState;
+            return (alignment ? util::IsAligned(chunk.virtualAddress, alignment) : true) && chunk.size > size && chunk.state == desiredState;
         })};
 
         if (chunk != chunks.end())
@@ -25,12 +27,12 @@ namespace skyline::gpu::vmm {
         return std::nullopt;
     }
 
-    u64 MemoryManager::InsertChunk(const ChunkDescriptor &newChunk) {
+    u64 GraphicsMemoryManager::InsertChunk(const ChunkDescriptor &newChunk) {
         auto chunkEnd{chunks.end()};
         for (auto chunk{chunks.begin()}; chunk != chunkEnd; chunk++) {
             if (chunk->CanContain(newChunk)) {
                 auto oldChunk{*chunk};
-                u64 newSize{newChunk.virtAddr - chunk->virtAddr};
+                u64 newSize{newChunk.virtualAddress - chunk->virtualAddress};
                 u64 extension{chunk->size - newSize - newChunk.size};
 
                 if (newSize == 0) {
@@ -41,16 +43,16 @@ namespace skyline::gpu::vmm {
                 }
 
                 if (extension)
-                    chunks.insert(std::next(chunk), ChunkDescriptor(newChunk.virtAddr + newChunk.size, extension, (oldChunk.state == ChunkState::Mapped) ? (oldChunk.cpuPtr + newSize + newChunk.size) : nullptr, oldChunk.state));
+                    chunks.insert(std::next(chunk), ChunkDescriptor(newChunk.virtualAddress + newChunk.size, extension, (oldChunk.state == ChunkState::Mapped) ? (oldChunk.cpuPtr + newSize + newChunk.size) : nullptr, oldChunk.state));
 
-                return newChunk.virtAddr;
-            } else if (chunk->virtAddr + chunk->size > newChunk.virtAddr) {
-                chunk->size = newChunk.virtAddr - chunk->virtAddr;
+                return newChunk.virtualAddress;
+            } else if (chunk->virtualAddress + chunk->size > newChunk.virtualAddress) {
+                chunk->size = newChunk.virtualAddress - chunk->virtualAddress;
 
                 // Deletes all chunks that are within the chunk being inserted and split the final one
                 auto tailChunk{std::next(chunk)};
                 while (tailChunk != chunkEnd) {
-                    if (tailChunk->virtAddr + tailChunk->size >= newChunk.virtAddr + newChunk.size)
+                    if (tailChunk->virtualAddress + tailChunk->size >= newChunk.virtualAddress + newChunk.size)
                         break;
 
                     tailChunk = chunks.erase(tailChunk);
@@ -61,8 +63,8 @@ namespace skyline::gpu::vmm {
                 if (tailChunk == chunkEnd)
                     break;
 
-                u64 chunkSliceOffset{newChunk.virtAddr + newChunk.size - tailChunk->virtAddr};
-                tailChunk->virtAddr += chunkSliceOffset;
+                u64 chunkSliceOffset{newChunk.virtualAddress + newChunk.size - tailChunk->virtualAddress};
+                tailChunk->virtualAddress += chunkSliceOffset;
                 tailChunk->size -= chunkSliceOffset;
                 if (tailChunk->state == ChunkState::Mapped)
                     tailChunk->cpuPtr += chunkSliceOffset;
@@ -74,19 +76,19 @@ namespace skyline::gpu::vmm {
                 else
                     chunks.insert(std::next(headChunk), newChunk);
 
-                return newChunk.virtAddr;
+                return newChunk.virtualAddress;
             }
         }
 
         throw exception("Failed to insert chunk into GPU address space!");
     }
 
-    u64 MemoryManager::ReserveSpace(u64 size, u64 alignment) {
-        size = util::AlignUp(size, constant::GpuPageSize);
+    u64 GraphicsMemoryManager::ReserveSpace(u64 size, u64 alignment) {
+        size = util::AlignUp(size, GpuPageSize);
 
-        std::unique_lock lock(vmmMutex);
+        std::unique_lock lock(mutex);
         auto newChunk{FindChunk(ChunkState::Unmapped, size, alignment)};
-        if (!newChunk)
+        if (!newChunk) [[unlikely]]
             return 0;
 
         auto chunk{*newChunk};
@@ -96,22 +98,22 @@ namespace skyline::gpu::vmm {
         return InsertChunk(chunk);
     }
 
-    u64 MemoryManager::ReserveFixed(u64 virtAddr, u64 size) {
-        if (!util::IsAligned(virtAddr, constant::GpuPageSize))
+    u64 GraphicsMemoryManager::ReserveFixed(u64 virtualAddress, u64 size) {
+        if (!util::IsAligned(virtualAddress, GpuPageSize)) [[unlikely]]
             return 0;
 
-        size = util::AlignUp(size, constant::GpuPageSize);
+        size = util::AlignUp(size, GpuPageSize);
 
-        std::unique_lock lock(vmmMutex);
-        return InsertChunk(ChunkDescriptor(virtAddr, size, nullptr, ChunkState::Reserved));
+        std::unique_lock lock(mutex);
+        return InsertChunk(ChunkDescriptor(virtualAddress, size, nullptr, ChunkState::Reserved));
     }
 
-    u64 MemoryManager::MapAllocate(u8 *cpuPtr, u64 size) {
-        size = util::AlignUp(size, constant::GpuPageSize);
+    u64 GraphicsMemoryManager::MapAllocate(u8 *cpuPtr, u64 size) {
+        size = util::AlignUp(size, GpuPageSize);
 
-        std::unique_lock lock(vmmMutex);
+        std::unique_lock lock(mutex);
         auto mappedChunk{FindChunk(ChunkState::Unmapped, size)};
-        if (!mappedChunk)
+        if (!mappedChunk) [[unlikely]]
             return 0;
 
         auto chunk{*mappedChunk};
@@ -122,23 +124,23 @@ namespace skyline::gpu::vmm {
         return InsertChunk(chunk);
     }
 
-    u64 MemoryManager::MapFixed(u64 virtAddr, u8 *cpuPtr, u64 size) {
-        if (!util::IsAligned(virtAddr, constant::GpuPageSize))
+    u64 GraphicsMemoryManager::MapFixed(u64 virtualAddress, u8 *cpuPtr, u64 size) {
+        if (!util::IsAligned(virtualAddress, GpuPageSize)) [[unlikely]]
             return 0;
 
-        size = util::AlignUp(size, constant::GpuPageSize);
+        size = util::AlignUp(size, GpuPageSize);
 
-        std::unique_lock lock(vmmMutex);
-        return InsertChunk(ChunkDescriptor(virtAddr, size, cpuPtr, ChunkState::Mapped));
+        std::unique_lock lock(mutex);
+        return InsertChunk(ChunkDescriptor(virtualAddress, size, cpuPtr, ChunkState::Mapped));
     }
 
-    bool MemoryManager::Unmap(u64 virtAddr, u64 size) {
-        if (!util::IsAligned(virtAddr, constant::GpuPageSize))
+    bool GraphicsMemoryManager::Unmap(u64 virtualAddress, u64 size) {
+        if (!util::IsAligned(virtualAddress, GpuPageSize)) [[unlikely]]
             return false;
 
         try {
-            std::unique_lock lock(vmmMutex);
-            InsertChunk(ChunkDescriptor(virtAddr, size, nullptr, ChunkState::Unmapped));
+            std::unique_lock lock(mutex);
+            InsertChunk(ChunkDescriptor(virtualAddress, size, nullptr, ChunkState::Unmapped));
         } catch (const std::exception &e) {
             return false;
         }
@@ -146,20 +148,20 @@ namespace skyline::gpu::vmm {
         return true;
     }
 
-    void MemoryManager::Read(u8 *destination, u64 virtAddr, u64 size) {
-        std::shared_lock lock(vmmMutex);
+    void GraphicsMemoryManager::Read(u8 *destination, u64 virtualAddress, u64 size) {
+        std::shared_lock lock(mutex);
 
-        auto chunk{std::upper_bound(chunks.begin(), chunks.end(), virtAddr, [](const u64 address, const ChunkDescriptor &chunk) -> bool {
-            return address < chunk.virtAddr;
+        auto chunk{std::upper_bound(chunks.begin(), chunks.end(), virtualAddress, [](const u64 address, const ChunkDescriptor &chunk) -> bool {
+            return address < chunk.virtualAddress;
         })};
 
         if (chunk == chunks.end() || chunk->state != ChunkState::Mapped)
-            throw exception("Failed to read region in GPU address space: Address: 0x{:X}, Size: 0x{:X}", virtAddr, size);
+            throw exception("Failed to read region in GPU address space: Address: 0x{:X}, Size: 0x{:X}", virtualAddress, size);
 
         chunk--;
 
         u64 initialSize{size};
-        u64 chunkOffset{virtAddr - chunk->virtAddr};
+        u64 chunkOffset{virtualAddress - chunk->virtualAddress};
         u8 *source{chunk->cpuPtr + chunkOffset};
         u64 sourceSize{std::min(chunk->size - chunkOffset, size)};
 
@@ -170,7 +172,7 @@ namespace skyline::gpu::vmm {
             size -= sourceSize;
             if (size) {
                 if (++chunk == chunks.end() || chunk->state != ChunkState::Mapped)
-                    throw exception("Failed to read region in GPU address space: Address: 0x{:X}, Size: 0x{:X}", virtAddr, size);
+                    throw exception("Failed to read region in GPU address space: Address: 0x{:X}, Size: 0x{:X}", virtualAddress, size);
 
                 source = chunk->cpuPtr;
                 sourceSize = std::min(chunk->size, size);
@@ -178,20 +180,20 @@ namespace skyline::gpu::vmm {
         }
     }
 
-    void MemoryManager::Write(u8 *source, u64 virtAddr, u64 size) {
-        std::shared_lock lock(vmmMutex);
+    void GraphicsMemoryManager::Write(u8 *source, u64 virtualAddress, u64 size) {
+        std::shared_lock lock(mutex);
 
-        auto chunk{std::upper_bound(chunks.begin(), chunks.end(), virtAddr, [](const u64 address, const ChunkDescriptor &chunk) -> bool {
-            return address < chunk.virtAddr;
+        auto chunk{std::upper_bound(chunks.begin(), chunks.end(), virtualAddress, [](const u64 address, const ChunkDescriptor &chunk) -> bool {
+            return address < chunk.virtualAddress;
         })};
 
         if (chunk == chunks.end() || chunk->state != ChunkState::Mapped)
-            throw exception("Failed to write region in GPU address space: Address: 0x{:X}, Size: 0x{:X}", virtAddr, size);
+            throw exception("Failed to write region in GPU address space: Address: 0x{:X}, Size: 0x{:X}", virtualAddress, size);
 
         chunk--;
 
         u64 initialSize{size};
-        u64 chunkOffset{virtAddr - chunk->virtAddr};
+        u64 chunkOffset{virtualAddress - chunk->virtualAddress};
         u8 *destination{chunk->cpuPtr + chunkOffset};
         u64 destinationSize{std::min(chunk->size - chunkOffset, size)};
 
@@ -202,7 +204,7 @@ namespace skyline::gpu::vmm {
             size -= destinationSize;
             if (size) {
                 if (++chunk == chunks.end() || chunk->state != ChunkState::Mapped)
-                    throw exception("Failed to write region in GPU address space: Address: 0x{:X}, Size: 0x{:X}", virtAddr, size);
+                    throw exception("Failed to write region in GPU address space: Address: 0x{:X}, Size: 0x{:X}", virtualAddress, size);
 
                 destination = chunk->cpuPtr;
                 destinationSize = std::min(chunk->size, size);
